Add AWS Copilot Formula
Copilot is a new CLI from the ECS Developer Experience team
to make developing, releasing and operating container apps on
AWS delightful.

This PR is to add Copilot to the AWS tap. Since our CLI is vended
in static binaries, we can skip the build step and use our pre-built
binaries which are produced as part of our release pipeline.

Right now we have binaries for MacOS and Intel Linux (ARM to come).

We also install shell autocompletion on install.

The config provider is a *tiny* bit awkward but I was able to make
it work :)

Thanks so much for hosting this tap!