Modify cbmc-viewer formula to depend on python and not python@3.7.
Depending on python@3.7 is a problem on M1.  Doing

```
brew uninstall cbmc-viewer
brew update-reset
brew install cbmc-viewer
```

on M1 generates the error message

```
python@3.7: The x86_64 architecture is required for this software.
Error: cbmc-viewer: An unsatisfied requirement failed this build.
```

One issue may be that there is no M1 bottle for python@3.7 and pip
needs to build 3.7 from scratch on M1.  But this error exists on M1
even with brew python 3.9 installed, so there should be no need for
installing 3.7.